agents/researcher/requirements.txt
strands-agents
a2a-sdk
fastapi
uvicorn